third_party/blink/web_tests/platform/generic/external/wpt/fetch/api/request/request-init-stream.any.sharedworker-expected.txt

This is a testharness.js-based test.
PASS Constructing a Request with a stream holds the original object.
PASS Constructing a Request with a stream on which getReader() is called
PASS Constructing a Request with a stream on which read() is called
PASS Constructing a Request with a stream on which read() and releaseLock() are called
PASS Constructing a Request with a Request on which body.getReader() is called
PASS Constructing a Request with a Request on which body.getReader().read() is called
PASS Constructing a Request with a Request on which read() and releaseLock() are called
PASS It is OK to omit .duplex when the body is null.
PASS It is OK to omit .duplex when the body is a string.
PASS It is OK to omit .duplex when the body is a Uint8Array.
PASS It is OK to omit .duplex when the body is a Blob.
FAIL It is error to omit .duplex when the body is a ReadableStream. assert_throws_js: function "() => new Request("...", { method, body })" did not throw
PASS It is OK to set .duplex = 'half' when the body is null.
PASS It is OK to set .duplex = 'half' when the body is a string.
PASS It is OK to set .duplex = 'half' when the body is a Uint8Array.
PASS It is OK to set .duplex = 'half' when the body is a Blob.
PASS It is OK to set .duplex = 'half' when the body is a ReadableStream.
FAIL It is error to set .duplex = 'full' when the body is null. assert_throws_js: function "() => new Request("...", { method, body, duplex })" did not throw
FAIL It is error to set .duplex = 'full' when the body is a string. assert_throws_js: function "() => new Request("...", { method, body, duplex })" did not throw
FAIL It is error to set .duplex = 'full' when the body is a Uint8Array. assert_throws_js: function "() => new Request("...", { method, body, duplex })" did not throw
FAIL It is error to set .duplex = 'full' when the body is a Blob. assert_throws_js: function "() => new Request("...", { method, body, duplex })" did not throw
FAIL It is error to set .duplex = 'full' when the body is a ReadableStream. assert_throws_js: function "() => new Request("...", { method, body, duplex })" did not throw
PASS It is OK to omit duplex when init.body is not given and input.body is given.
Harness: the test ran to completion.
